// src/Akeneo/Pim/Automation/FranklinInsights/Infrastructure/Symfony/Resources/public/react/domain/model/front-attribute-mapping-status.enum.ts
/*
 * This file is part of the Akeneo PIM Enterprise Edition.
 *
 * (c) 2019 Akeneo SAS (http://www.akeneo.com)
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */

export enum FrontAttributeMappingStatus {
  PENDING = 'pending',
  MAPPED = 'mapped',
  SUGGESTION_APPLIED = 'suggestion_applied'
}